+++ cogs/music/classes/music.py
@@ -5,13 +5,13 @@ instances of the music player.
 import logging
 
 from discord.ext import commands
-from .player import Player
 
-logger = logging.getLogger("musicplayer")
+from .player import Player
 
 class Music:
     def __init__(self) -> None:
         self.__players  = {}
+        self.logger: logging.Logger = logging.getLogger("music")
 
     def create_player(self, ctx: commands.Context) -> Player:
         """Create player instance"""
@@ -20,7 +20,7 @@ class Music:
 
         player = Player(ctx)
         self.__players[ctx.guild.id] = player
-        logger.info(f"Created Player instance (Guild ID: {player.channel.guild.id})")
+        self.logger.info(f"Created Player instance (Guild ID: {player.channel.guild.id})")
         return player
 
     def get_player(self, player_id: int) -> Player:
@@ -32,6 +32,6 @@ class Music:
         player = self.get_player(player_id)
 
         if player:
-            logger.info(f"Destroyed Player instance (Guild ID: {player.channel.guild.id})")
+            self.logger.info(f"Destroyed Player instance (Guild ID: {player.channel.guild.id})")
             player.loop.create_task(player.stop())
             self.__players.pop(player_id)
