+++ cogs/music/classes/ui/components/__init__.py
@@ -1,0 +1,5 @@
+from .next_button import NextButton
+from .pause_button import PauseButton
+from .play_button import PlayButton
+from .prev_button import PrevButton
+from .stop_button import StopButton
