+++ cogs/music/colors.py
@@ -1,0 +1,11 @@
+"""Define color variables here"""
+import discord
+
+#define colors
+green       = discord.Colour.green()
+red         = discord.Colour.red()
+gold        = discord.Colour.gold()
+purple      = discord.Colour.purple()
+dark_blue   = discord.Colour.dark_blue()
+blue        = discord.Colour.blue()
+pink        = discord.Colour.from_rgb(253, 0, 97)
