+++ core/logsettings.py
@@ -42,7 +42,27 @@ LOGGING_CONFIG = {
             "level"     : "INFO",
             "propagate" : False
         },
-        "musicplayer"   : {
+        "music"   : {
+            "handlers"  : ["console"],
+            "level"     : "DEBUG",
+            "propagate" : False
+        },
+        "music.commands"   : {
+            "handlers"  : ["console"],
+            "level"     : "DEBUG",
+            "propagate" : False
+        },
+        "music.song"   : {
+            "handlers"  : ["console"],
+            "level"     : "DEBUG",
+            "propagate" : False
+        },
+        "music.player"   : {
+            "handlers"  : ["console"],
+            "level"     : "DEBUG",
+            "propagate" : False
+        },
+        "music.player.ui"   : {
             "handlers"  : ["console"],
             "level"     : "DEBUG",
             "propagate" : False
