+++ version.py
@@ -3,7 +3,7 @@ This file conains the version information of the bot.
 """
 
 APP_NAME = "Yae Miko"
-__version__ = "1.5.0"
+__version__ = "1.6.0"
 __author__ = "cxrvxxxx"
 __license__ = "GPL-3.0"
 __description__ = "A feature-packed Discord bot using discord.py 2.0"
